Fix for #108: Order read tickets by date
A couple notes on this: 
 * This option is triggered by 'rss.config.feedgrouping', which is also used
for unread items. Should we have a distinct config entry?

 * I've disabled collapsing of read feeds when this
option is set, because you could have several instances of a given feed
(which one should Gregarius collapse? Each one?)

 * How many read items should be displayed? I've arbitrarly set this to be
(rss.output.numitemsonpage - unread count) when there are unread items, and
to 50 when only read items are shown.



git-svn-id: http://svn.gregarius.net/svn/trunk@757 d826556c-87f9-0310-90b4-9554b5b11594