Better protection against cookie theft: when a user logs in, his IP address subnet is stored in a list of valid subnets. Each time the cookie hash is checked, we also check whether the user's IP address matches a registerd subnet: if the IP isn't in the list, the cookie isn't validated.
Logging out currently also removes the user's IP from the list. This might not 
be the perfect solution as someone browsing with two different brwosers or
from two computers, and logging out on either one, also invalidates the cookie 
on the other browser.

On the other hand we don't want the list of subnets to grow, with the risk
of being polluted. (Imagine the case of a user browsing from a public
terminal, in an Internet Cafe, say: we definitely want that IP to be removed 
from the list of valid IPs when the user logs out)



git-svn-id: http://svn.gregarius.net/svn/trunk@1298 d826556c-87f9-0310-90b4-9554b5b11594